% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_foundry10.R
\name{foundry10_palette}
\alias{foundry10_palette}
\title{Return a named vector of foundry10 colors by palette group}
\usage{
foundry10_palette(palette = "primary")
}
\arguments{
\item{palette}{Character. One of \code{"primary"}, \code{"secondary"},
\code{"tints_orange"}, \code{"tints_cyan"}, \code{"gray"}, or \code{"all"}.}
}
\value{
A named character vector of hex codes.
}
\description{
Return a named vector of foundry10 colors by palette group
}
\examples{
foundry10_palette("primary")
foundry10_palette("gray")
}
